Adv loop ops finished and tested with osexp, starting with unit tests
